% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_biodepth.R
\docType{data}
\name{all_biodepth}
\alias{all_biodepth}
\title{Biodepth Data}
\description{
Data from the pan-European Biodepth grassland diversity manipulation
}
\references{
Spehn, E. M., A. Hector, J. Joshi, M. Scherer-Lorenzen, B. Schmid,
E. Bazeley-White, C. Beierkuhnlein, M. C. Caldeira,
M. Diemer, P. G. Dimitrakopoulos, J. A. Finn, H. Freitas, P. S. Giller,
J. Good, R. Harris, P. Hogberg, K. Huss-Danell, A. Jumpponen, J. Koricheva,
P. W. Leadley, M. Loreau, A. Minns, C. P. H. Mulder, G. O'Donovan, S. J. Otway,
C. Palmborg, J. S. Pereira, A. B. Pfisterer, A. Prinz, D. J. Read, E. D. Schulze,
A. S. Siamantziouras, A. C. Terry, A. Y. Troumbis, F. I. Woodward, S. Yachi, and J. H. Lawton. 2005.
Ecosystem effects of biodiversity manipulations in European grasslands.
Ecological Monographs 75:37-63.
}
\author{
Jarrett Byrnes
}
\keyword{data}
