% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_num_func.R
\name{cor_dist}
\alias{cor_dist}
\title{cor_dist}
\usage{
cor_dist(adf)
}
\arguments{
\item{adf}{A \code{data.frame} or \code{matrix} of functions}
}
\value{
A matrix
}
\description{
Takes a data frame of functions and calculates the correlation-based
distance between functions.
}
