% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_num_func.R
\name{dmean}
\alias{dmean}
\title{dmean}
\usage{
dmean(adf_raw, D)
}
\arguments{
\item{D}{A distance matrix describing dissimilarity between functions.}

\item{dat}{A data frame with functions in columns and rows as replicates}
}
\value{

}
\description{
Calculates the average distance between functions for one or an entire
assemblage of replicates
}
\references{
Chao, A., Chiu, C.-H., Villéger, S., Sun, I.-F., Thorn, S., Lin, Y.-C., 
Chiang, J.-M. and Sherwin, W. B. 2019. An attribute-diversity approach to 
functional diversity, functional beta diversity, and related (dis)similarity 
measures. Ecological Monographs. 89: e01343.
}
