% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_num_func.R
\name{dmin}
\alias{dmin}
\title{dmin}
\usage{
dmin(D)
}
\arguments{
\item{D}{A distance matrix describing dissimilarity between functions.}
}
\value{
A numeric
}
\description{
Calculates the minimum non-zero value of a distance matrix
}
