% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_div.R
\name{eff_div}
\alias{eff_div}
\title{eff_div}
\usage{
eff_div(df, q = 1)
}
\arguments{
\item{df}{A data frame of functions}

\item{q}{Order of the diversity measure. Defaults to the 
Shannon case where q = 1. For Simpson, q=2.}
}
\value{
Returns a vector.
}
\description{
\code{eff_div} effective number of functions
}
\details{
Takes a data frame, variable names, and type of index and returns
 the effective number of functions - i.e., the equivalent number of 
 functions if all were performing at the same level. See Jost 2006 and 2010.
 Internally, uses \code{\link[vegetarian]{d}}
}
\author{
Jarrett Byrnes.
}
