% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_num_func.R
\name{eff_num_func_d}
\alias{eff_num_func_d}
\title{eff_num_func_d}
\usage{
eff_num_func_d(adf_freq, q = 1, D, tau = NULL)
}
\arguments{
\item{adf_freq}{A data frame of functional "frequencies" - i.e. f_i/sum(f_i)}

\item{q}{Order of hill number used for index. Defaults to q=1, as in Shannon Diversity}

\item{D}{A distance matrix describing dissimilarity between functions.}

\item{tau}{A cutoff for degree of dissimilarity under which functions are considered
to be different. If tau is the minimum non-zero value of D, all functions are different.
if tau is the maximum value of D are greater, all functions are considered the same.}
}
\value{
A vector of effective number of functions
}
\description{
eff_num_func_d
}
\references{
Chao, A., Chiu, C.-H., Villéger, S., Sun, I.-F., Thorn, S., Lin, Y.-C.,
Chiang, J.-M. and Sherwin, W. B. 2019. An attribute-diversity approach to
functional diversity, functional beta diversity, and related (dis)similarity
measures. Ecological Monographs. 89: e01343.
}
