% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_num_func.R
\name{eff_num_func_d_onerow}
\alias{eff_num_func_d_onerow}
\title{eff_num_func_d_onerow}
\usage{
eff_num_func_d_onerow(arow_freq, D, tau, q)
}
\arguments{
\item{arow_freq}{One replicate sample of different functions (a single numeric vector)}

\item{D}{A distance matrix describing dissimilarity between functions.}

\item{tau}{A cutoff for degree of dissimilarity under which functions are considered
to be different. If tau is the minimum non-zero value of D, all functions are different.
if tau is the maximum value of D are greater, all functions are considered the same.}

\item{q}{Order of hill number used for index. Defaults to q=1, as in Shannon Diversity}
}
\value{
A single value of effective number of functions
}
\description{
eff_num_func_d_onerow
}
\references{
Byrnes, J. E. K., Roger, F. and Bagchi, R. 2022. Understandable 
Multifunctionality Measures Using Hill Numbers. bioRxiv. 
2022.03.17.484802. https://doi.org/10.1101/2022.03.17.484802

Chao, A., Chiu, C.-H., Villéger, S., Sun, I.-F., Thorn, S., Lin, Y.-C.,
Chiang, J.-M. and Sherwin, W. B. 2019. An attribute-diversity approach to
functional diversity, functional beta diversity, and related (dis)similarity
measures. Ecological Monographs. 89: e01343.
}
