% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcEven.R
\name{even_fact}
\alias{even_fact}
\title{even_fact}
\usage{
even_fact(df, q = 1)
}
\arguments{
\item{df}{A data frame of functions}

\item{q}{Order of the diversity measure. Defaults to the 
Shannon case where q = 1. For Simpson, q=2.}
}
\value{
Returns a vector.
}
\description{
\code{even_fact} evenness factor across functions
}
\details{
Takes a data frame, variable names, and type of index and gets 
the evenness factor for the collection of values supplied as defined in 
Jost 2010
}
\author{
Jarrett Byrnes.
}
