% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcDiv.R
\name{funcDiv}
\alias{funcDiv}
\title{funcDiv}
\usage{
funcDiv(data, vars, type = "Simpson")
}
\arguments{
\item{data}{A vector of measurements of a function.}

\item{vars}{Name of function variables}

\item{type}{Diversity index to be used. Options are "Simpson" and "Shannon"}
}
\value{
Returns a vector.
}
\description{
\code{funcDiv} effective number of functions
}
\details{
Takes a data frame, variable names, and type of index and returns
 the effective number of functions - i.e., the equivalent number of 
 functions if all were performing at the same level. See Jost 2006 and 2010.
}
\author{
Jarrett Byrnes.
}

