% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcEven.R
\name{funcEven}
\alias{funcEven}
\title{funcEven}
\usage{
funcEven(data, vars, q = 1)
}
\arguments{
\item{data}{A vector of measurements of a function.}

\item{vars}{Name of function variables}

\item{q}{Order of the diversity measure. Defaults to the 
Shannon case where q = 1. For Simpson, q=2.}
}
\value{
Returns a vector.
}
\description{
\code{funcEven} evenness factor across functions
}
\details{
Takes a data frame, variable names, and type of index and gets 
the evenness factor for the collection of values supplied as defined in 
Jost  2010
}
\author{
Jarrett Byrnes.
}
\references{
Jost, L. 2006. Entropy and diversity. Oikos 113(2): 363-375.
}

