% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_num_func.R
\name{dmean}
\alias{dmean}
\title{dmean}
\usage{
dmean(adf_raw, D)
}
\arguments{
\item{adf_raw}{A data frame frame with functions in columns and rows as replicates}

\item{D}{A distance matrix describing dissimilarity between functions.}
}
\value{
Single numeric of weighted average of distance matrix
}
\description{
Calculates the average distance between functions for one or an entire
assemblage of replicates
}
\references{
Byrnes, J. E. K., Roger, F. and Bagchi, R. 2022. Understandable 
Multifunctionality Measures Using Hill Numbers. bioRxiv. 
2022.03.17.484802. https://doi.org/10.1101/2022.03.17.484802

Chao, A., Chiu, C.-H., Villéger, S., Sun, I.-F., Thorn, S., Lin, Y.-C.,
Chiang, J.-M. and Sherwin, W. B. 2019. An attribute-diversity approach to
functional diversity, functional beta diversity, and related (dis)similarity
measures. Ecological Monographs. 89: e01343.
}
