% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duffy_2003.R
\docType{data}
\name{duffy_2003}
\alias{duffy_2003}
\title{Seagrass Mesocosm Data}
\description{
Data from the a seagrass grazer diversity manipulation at the Virginia Institute of Marine Sciences. From Duffy et al. 2003 Ecology Letters.
}
\references{
Duffy, J. E., J. P. Richardson, and E. A. Canuel. 2003. Grazer diversity effects on ecosystem functioning in seagrass beds. Ecology letters 6:637-645.
}
\author{
Jarrett Byrnes
}
\keyword{data}
