% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcDiv.R
\name{eff_div}
\alias{eff_div}
\title{eff_div}
\usage{
eff_div(df, type = "Simpson")
}
\arguments{
\item{df}{A data frame of functions}

\item{type}{Diversity index to be used. Options are "Simpson" and "Shannon"}
}
\value{
Returns a vector.
}
\description{
\code{eff_div} effective number of functions
}
\details{
Takes a data frame, variable names, and type of index and returns
 the effective number of functions - i.e., the equivalent number of 
 functions if all were performing at the same level. See Jost 2006 and 2010.
}
\author{
Jarrett Byrnes.
}

