% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcEven.R
\name{funcEven}
\alias{funcEven}
\title{funcEven}
\usage{
funcEven(data, vars, type = "Simpson")
}
\arguments{
\item{data}{A vector of measurements of a function.}

\item{vars}{Name of function variables}

\item{type}{Diversity index to be used. Options are "Simpson" and "Shannon"}
}
\value{
Returns a vector.
}
\description{
\code{funcEven} evenness factor across functions
}
\details{
Takes a data frame, variable names, and type of index and gets 
the evenness factor for the collection of values supplied as defined in 
Jost 2010
}
\author{
Jarrett Byrnes.
}

