% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMF.R
\name{getMF}
\alias{getMF}
\title{getMF}
\usage{
getMF(data, vars, standardizeFunction = standardizeUnitScale, q = 1)
}
\arguments{
\item{data}{A vector of measurements of a function.}

\item{vars}{Name of function variables}

\item{standardizeFunction}{A function to standardize each individual 
function to the same scale, such as \code{standardizeUnitScale} or 
\code{standardizeZScore}}

\item{q}{Order of the diversity measure. Defaults to the 
Shannon case where q = 1. For Simpson, q=2.}
}
\value{
Returns a vector.
}
\description{
An entropy-based multifunctionality index.
\code{getMF} get multifunctionality index defined by function and evenness
}
\details{
Takes a data frame, variable names, a standardizing function, and a type of diversity index and returns an index of multifunctionality. Here MF = MF_a * MF_e where MF_a is the average level of functionality across functions sampled and MF_e is the evenness factor from \code{funcEven}.
}
\author{
Jarrett Byrnes.
}

