% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qw.R
\name{qw}
\alias{qw}
\title{Quote Words}
\usage{
qw(...)
}
\arguments{
\item{...}{Any unquoted strings}
}
\value{
A vector
}
\description{
\code{qw} Takes an unquoted vector and adds quotes to it like the qw function in perl.
}
\details{
This is a helper function for data processing. Honestly, I use qw all the time
in other languages, and wanted a version for R.
}
\examples{
c("a", "b")

qw(a, b)



# qw - a helper function that we
# will use later to deal with strings
# analagous to qw in PERL
}
\author{
Jarrett Byrnes.
}
