% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeZScore.R
\name{standardizeZScore}
\alias{standardizeZScore}
\title{standardizeZScore}
\usage{
standardizeZScore(afun)
}
\arguments{
\item{afun}{A vector of measurements of a function.}
}
\value{
Returns a z-standardized vector.
}
\description{
\code{standardizeZScore} Z-standardizes a vector.
}
\details{
Centers a vector and divides it by its standard deviation.
}
\author{
Jarrett Byrnes.
}
